管理员更新 菜单列表 新增菜单

+++ routes/web.php
@@ -42,16 +42,34 @@ Route::namespace('Admin')->prefix('admin')->middleware('admin.auth')->group(func
 
     Route::prefix('index')->group(function () {
         // 后台首页
-        Route::get('index', 'IndexController@index');
+        Route::get('/index', 'IndexController@index');
     });
 
     //系统管理
     Route::prefix('auth')->group(function (){
         // 管理员列表
         Route::get('/users', 'AuthController@users')->name('admin.auth.users.index');
+        //新增管理员视图
         Route::get('/users/create', 'AuthController@createUser')->name('admin.auth.users.create');
-        Route::post('/users', 'AuthController@storeUser');
+        //创建管理员
+        Route::post('/users', 'AuthController@storeUser')->name('admin.auth.users.store');
+        //编辑管理员视图
         Route::get('/users/{user}/edit', 'AuthController@editUser')->name('admin.auth.users.edit');
-        Route::delete('/users/{user}', 'AuthController@destroyUser');
+        //更新管理员
+        Route::patch('/users/{user}', 'AuthController@updateUser')->name('admin.auth.users.update');
+        //删除管理员
+        Route::delete('/users/{user}', 'AuthController@destroyUser')->name('admin.auth.users.delete');
+    });
+
+    //系统管理
+    Route::prefix('nav')->group(function (){
+        //菜单列表
+        Route::get('/index','NavController@index')->name('admin.nav.index');
+        //新增菜单
+        Route::post('/store','NavController@store')->name('admin.nav.store');
+        //编辑菜单视图
+        Route::get('/{nav}/edit', 'NavController@edit')->name('admin.nav.edit');
+        //删除菜单
+        Route::post('/destroy/{nav}', 'NavController@destroy')->name('admin.nav.destroy');
     });
 });
